+++ src/RoadRegistry/RoadSegment/Events/V1/ValueObjects/RoadSegmentSurfaceAttributes.cs
@@ -4,7 +4,7 @@ public class RoadSegmentSurfaceAttributes
 {
     public required int AsOfGeometryVersion { get; set; }
     public required int AttributeId { get; set; }
-    public required decimal FromPosition { get; set; }
-    public required decimal ToPosition { get; set; }
+    public required double FromPosition { get; set; }
+    public required double ToPosition { get; set; }
     public required string Type { get; set; }
 }
